going to need to do some major clean up now but login and users sessions are now working for all of the current fucntions.

+++ houser/Data/PropertyListDB.cs
@@ -17,28 +17,22 @@ namespace houser.Data
         /// <summary>
         /// Remove property by accountnumber and listid
         /// </summary>
-        public static void RemoveFromFromList(string accountNumber, int list)
+        public static void RemoveFromFromList(string accountNumber, int list, int userID)
         {
             SqlHelper.ExecuteNonQuery(CONNECTIONSTRING, CommandType.Text,
-                "DELETE FROM PropertyList WHERE AccountNumber = @AccountNumber AND ListID = @ListID",
+                "DELETE FROM PropertyList WHERE AccountNumber = @AccountNumber AND ListID = @ListID AND UserID = @UserID",
                 new SqlParameter("@ListID", list),
-                new SqlParameter("@AccountNumber", accountNumber));
+                new SqlParameter("@AccountNumber", accountNumber),
+                new SqlParameter("@UserID", userID));
         }
 
-        public static void UpdatePropertyList(string accountNumber, int list)
+        public static void AddPropertyToList(string accountNumber, int list, int userID)
         {
             SqlHelper.ExecuteNonQuery(CONNECTIONSTRING, CommandType.Text,
-                "UPDATE PropertyList SET ListID = @ListID WHERE AccountNumber = @AccountNumber",
+                "INSERT INTO PropertyList (ListID, AccountNumber, UserID) VALUES (@ListID, @AccountNumber, @UserID)",
                 new SqlParameter("@ListID", list),
-                new SqlParameter("@AccountNumber", accountNumber));
-        }
-
-        public static void AddPropertyToList(string accountNumber, int list)
-        {
-            SqlHelper.ExecuteNonQuery(CONNECTIONSTRING, CommandType.Text,
-                "INSERT INTO PropertyList (ListID, AccountNumber) VALUES (@ListID, @AccountNumber)",
-                new SqlParameter("@ListID", list),
-                new SqlParameter("@AccountNumber", accountNumber));
+                new SqlParameter("@AccountNumber", accountNumber),
+                new SqlParameter("@UserID", userID));
         }
 
         public static bool PropertyListExist(string accountNumber, int list)
@@ -49,12 +43,13 @@ namespace houser.Data
             return result != null;
         }
 
-        public static bool PropertyListExistByAccountAndList(string accountNumber, int list)
+        public static bool PropertyListExistByAccountAndList(string accountNumber, int list, int userID)
         {
             var result = SqlHelper.ExecuteScalar(CONNECTIONSTRING, CommandType.Text,
-                "SELECT ListID FROM PropertyList WHERE ListID = @ListID AND AccountNumber = @AccountNumber",
+                "SELECT ListID FROM PropertyList WHERE ListID = @ListID AND AccountNumber = @AccountNumber AND UserID = @UserID",
                 new SqlParameter("@AccountNumber", accountNumber),
-                new SqlParameter("@ListID", list));
+                new SqlParameter("@ListID", list),
+                new SqlParameter("@UserID", userID));
             return result != null;
         }
         #endregion
